+++ src/dapps/MakerDai/MakerCDP/index.js
@@ -1,0 +1,3 @@
+import MakerCDP from '@/dapps/MakerDai/MakerCDP/MakerCDP';
+
+export default MakerCDP;